try to fix events by waiting for umami to become available

+++ bundle/init.js
@@ -17,12 +17,23 @@ domready(() => {
   });
 
   // complicated logic to fire events and retry if the page reloaded in the interim
-  async function fireEvents() {
+  function fireEvents() {
     const events = get("eventCache", [], "sessionStorage");
-    try {
-      await Promise.all(events.map((event) => umami(element.dataset.event)));
-    } catch (e) {}
-    unlink("eventCache", "sessionStorage");
+    let count = 0;
+    const interval = setInterval(function () {
+      if (!window.umami) {
+        count++;
+        // give up after 15 seconds.
+        if (count > 15) {
+          clearInterval(interval);
+          unlink("eventCache", "sessionStorage");
+        }
+        return;
+      }
+      clearInterval(interval);
+      Promise.all(events.map((event) => umami(element.dataset.event)));
+      unlink("eventCache", "sessionStorage");
+    }, 1000);
   }
   document.querySelectorAll("[data-event]").forEach((element) => {
     element.addEventListener("click", async (e) => {
